Hide video view if the participant does not have publishing permissions
If the participant can not send media or screenshare it is now hidden
from the call view.

If there is a single publisher in the call then that publisher would see
the empty call view message, as the rest of participants would be
hidden. To solve that a dummy participant that just shows a count of all
the subscriber-only participants in the call is added to the call view
(if there is at least one subscriber-only participant).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>