Use peer data to hide listeners if participant data is not available
Guests can not fetch the full participant list, so the participant store
is empty when the current user is a guest. However, guests can fetch the
peer data, so now that data is used when the participant data is not
available to find out if another participant has publishing permissions
or not and hide her if needed in the call view.

The peer data is extended with "publishing permissions" and it is also
updated on "participantListChanged" events emitted by the signaling, as
until now it was updated only if it was required by a Video element but
it was not available.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>