Fix requested call flags set in the store instead of the actual ones
The requested call flags can be different of the actual call flags used
if the audio or video devices are not available. This is not a big
issue, as the flags will be updated with the proper value as soon as the
participant list is requested again. Nevertheless, it is better to set
the right ones :-)

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>