Restore media state only when joining a call
The previous media state was restored when the local media was started.
Until now that happened only when starting a call, but the media will
need to be started during a call too if publishing permissions are
granted. In that case the media state should not be restored and the
media must start always disabled, as granting the publishing permissions
is done by other users, not by the one that will be publishing.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>