Add UI feedback when local participant is not connected
When the connection with other participants is lost those participants
are shown dimmed and with a loading spinner. However, when the MCU is
used that only covers receiving audio and video from those participants.
Other participants can not hear or see the local participant until the
sender peer object has connected. Therefore now the local video / avatar
is dimmed and shows a loading spinner like done with the remote
participants when the sender peer object is not connected.

Note that when the MCU is used the sender participant will always
reconnect, so there is no separate "isLoading" property like the one
used for remote participants (which, moreover, is needed only when the
MCU is not used).

When the MCU is not used the connections are peer to peer between
partipants, so the already existing dimming and loading spinner on the
remote participants views already show whether those participants can
hear and see the local participant or not.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>